requirements.txt
django==5.1.2

django-cors-headers==3.13.0
cloudinary
 
django-environ
djangorestframework
djoser==2.1.0

django-rest-auth==0.9.5
django-allauth==0.52.0
django-ckeditor==6.1.0
dj-rest-auth==3.0.0

psycopg2-binary
argon2-cffi==21.1.0 
PyJWT==2.6.0
django-storages==1.12.2 
markdown==3.3.4
gunicorn==20.1.0
whitenoise==5.3.0
requests==2.26.0
django-anymail==8.6
mailjet-rest==1.3.4
Pillow
python-decouple==3.8
decouple
environ

